An Alpha Implementation of Tuning (#1)
Restructure dexed so midi_to_log_freq is in a data structure
owned by the PluginProcessor; have a default which is a table
into the values prior to this change; port in a copy of the
Surge SCL and KBM implementation; make it so you can load an
SCL and it retunes.

Still a million and one things to do. But this is a
testable starting alpha.