Update get-juce.sh to use the latest JUCE version
The linux nightly build is broken on Ubuntu due to a JUCE Makefile error; the correct std version is not being set. 
Error: `unordered_map’ in namespace ‘std’ does not name a template type` which indicates std=c++17 compiler flag is not being correctly set. The latest version of JUCE fixes this.